getAllCustomer -done

+++ src/main/java/lk/riyapola/riyapola/repo/CustomerRepo.java
@@ -3,5 +3,5 @@ package lk.riyapola.riyapola.repo;
 import lk.riyapola.riyapola.entity.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepo extends JpaRepository<Customer,Long> {
+public interface CustomerRepo extends JpaRepository<Customer,Integer> {
 }
